+++ providers/implementations/ciphers/cipher_aes_xts.c
@@ -8,7 +8,7 @@
  */
 
 #include "cipher_aes_xts.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 #include "internal/providercommonerr.h"
 
 /* TODO (3.0) Figure out what flags need to be set */
