+++ providers/implementations/ciphers/cipher_aria.c
@@ -10,7 +10,7 @@
 /* Dispatch functions for ARIA cipher modes ecb, cbc, ofb, cfb, ctr */
 
 #include "cipher_aria.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 
 static OSSL_OP_cipher_freectx_fn aria_freectx;
 static OSSL_OP_cipher_dupctx_fn aria_dupctx;
