+++ providers/implementations/ciphers/cipher_tdes.c
@@ -10,7 +10,7 @@
 #include "prov/ciphercommon.h"
 #include "cipher_tdes.h"
 #include "crypto/rand.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 #include "internal/providercommonerr.h"
 
 void *tdes_newctx(void *provctx, int mode, size_t kbits, size_t blkbits,
