+++ providers/implementations/digests/mdc2_prov.c
@@ -13,7 +13,7 @@
 #include <openssl/core_names.h>
 #include <openssl/err.h>
 #include "prov/digestcommon.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 #include "internal/providercommonerr.h"
 
 static OSSL_OP_digest_set_ctx_params_fn mdc2_set_ctx_params;
