+++ providers/implementations/kdfs/kbkdf.c
@@ -37,7 +37,7 @@
 #include "internal/cryptlib.h"
 #include "crypto/evp.h"
 #include "internal/numbers.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 #include "internal/provider_ctx.h"
 #include "internal/provider_util.h"
 #include "internal/providercommonerr.h"
