+++ providers/implementations/kdfs/sskdf.c
@@ -47,7 +47,7 @@
 #include "crypto/evp.h"
 #include "internal/provider_ctx.h"
 #include "internal/providercommonerr.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 #include "internal/provider_util.h"
 
 typedef struct {
