+++ providers/implementations/macs/gmac_prov.c
@@ -16,7 +16,7 @@
 #include <openssl/err.h>
 
 #include "internal/providercommonerr.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 #include "internal/provider_ctx.h"
 #include "internal/provider_util.h"
