+++ providers/implementations/macs/hmac_prov.c
@@ -14,7 +14,7 @@
 #include <openssl/evp.h>
 #include <openssl/hmac.h>
 
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 #include "internal/provider_ctx.h"
 #include "internal/provider_util.h"
