+++ providers/implementations/macs/siphash_prov.c
@@ -23,7 +23,7 @@
 #include "../../../crypto/siphash/siphash_local.h"
 
 #include "internal/providercommonerr.h"
-#include "internal/provider_algs.h"
+#include "prov/implementations.h"
 
 /*
  * Forward declaration of everything implemented here.  This is not strictly
